==> src/notebooks/00_parse_and_clean_data.ipynb <==
##### This notebook contains the preprocessing to the Raw Amazon Review and Meta data
##### The raw data files ca ne found [here](https://cseweb.ucsd.edu/~jmcauley/datasets/amazon_v2/)
##### The output is saved as reviews_full.csv and filtered_data.csv files and used as input for [next stage](01_data_preprocessing.ipynb)

import os
import json
import gzip
import pandas as pd
from urllib.request import urlopen

### load the meta data

data = []
with gzip.open('meta_Software.json.gz') as f:
    for l in f:
        data.append(json.loads(l.strip()))

# total length of list, this number equals total number of products
print(len(data))

# first row of the list
print(data[0])

# convert list into pandas dataframe

df = pd.DataFrame.from_dict(data)

print(len(df))

### remove rows with unformatted title (i.e. some 'title' may still contain html style content)

df3 = df.fillna('')
df4 = df3[df3.title.str.contains('getTime')] # unformatted rows
df5 = df3[~df3.title.str.contains('getTime')] # filter those unformatted rows
print(len(df4))
print(len(df5))

df5.head()

empty_categories = df5['category'].apply(lambda x: len(x) == 0).sum()
software_categories = df5['category'].apply(lambda x: 'Software' in x).sum()

print("Number of empty categories:", empty_categories)
print("Number of categories labeled as 'Software':", software_categories)


empty_prices = df5['price'].isna().sum()
software_prices = df5['price'].apply(lambda x: 'Software' in str(x)).sum()

print("Number of empty prices:", empty_prices)
print("Number of prices labeled as 'Software':", software_prices)


duplicate_asins = df5.duplicated(subset='asin').sum()
print("Number of duplicate ASINs:", duplicate_asins)


empty_descriptions = df5['description'].apply(lambda x: len(x) == 0).sum()
print("Number of empty descriptions:", empty_descriptions)


software_categories = df5['main_cat'].apply(lambda x: 'Software' in x).sum()
print("Number of categories labeled as 'Software':", software_categories)


non_div_date = df5[df5['date'] != '</div>'].shape[0]
print("Number of rows with date not equal to '</div>':", non_div_date)


filtered_df = df5[df5['main_cat'].apply(lambda x: 'Software' in x)]
filtered_df = filtered_df.drop_duplicates(subset='asin')
filtered_df = filtered_df[filtered_df['category'].apply(lambda x: 'Software' in x)]
filtered_df = filtered_df.drop(['imageURL', 'imageURLHighRes'], axis=1)

# Print the resulting DataFrame
print(filtered_df.shape)
filtered_df.head()


non_empty_also_buy = filtered_df['also_buy'].apply(lambda x: len(x) > 0).sum()
print("Number of rows with non-empty 'also_buy':", non_empty_also_buy)


non_empty_similar_item = filtered_df['similar_item'].apply(lambda x: len(x) > 0).sum()
print("Number of rows with non-empty 'similar_item':", non_empty_similar_item)


non_empty_similar_item = filtered_df['asin'].apply(lambda x: len(x) > 0).sum()
print("Number of rows with non-empty 'asin':", non_empty_similar_item)


filtered_df.to_csv('filtered_data.csv', index=False)


#load the data
data = []
with gzip.open('Software.json.gz') as f:
    for l in f:
        data.append(json.loads(l.strip()))

reviews = pd.DataFrame.from_dict(data)


reviews.head()

reviews = reviews.drop(['unixReviewTime', 'image','style'], axis=1)
filtered_reviews = reviews[reviews['asin'].isin(filtered_df['asin'])]

filtered_reviews.shape
filtered_reviews.head()

filtered_reviews.to_csv('reviews_full.csv', index=False)
