==> src/notebooks/01_data_preprocessing.ipynb <==
## Data Pre-processing
### In this notebook we processed load and clean the Amazon review and software meta data.

The preprocessing includes the following steps:
1. Remove stop words and html tags from text fields.
2. Generating license fee, maintenance fee and implementation fee, based on software category information.

To avoid duplication of code, we have this pre-processing task implemented in [here](../data_preprocessing.py)

import os
import pandas as pd
import numpy as np
import seaborn as sns
os.chdir('../')


meta_data_path = "../../external_data/filtered_metadata.csv"
review_data_path = "../../external_data/reviews_full.csv"

meta_data = pd.read_csv(meta_data_path)
review_data = pd.read_csv(review_data_path)

meta_data.head()

meta_data.info()

meta_data.also_view.value_counts()

### review data

review_data.head()

review_data.info()

### Pre-processing

from data_preprocessing import PreProcessor

obj = PreProcessor(meta_data_path ,review_data_path)
review_data , software_data = obj.main()
review_data.to_csv("../data/reviews.csv")
software_data.to_csv("../data/softwares.csv")

### reading output of data prepocessing 

reviews = pd.read_csv("../data/reviews.csv")
softwares = pd.read_csv("../data/softwares.csv")
review_meta = reviews.merge(softwares, on='asin', how='inner')

review_meta.head()

review_meta.columns

review_meta.info()

review_meta.describe()
