==> src/notebooks/03_LLM_encoding.ipynb <==
import pandas as pd
import nltk
from sentence_transformers import SentenceTransformer
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.cluster import KMeans
from transformers import pipeline
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity
import numpy as np

##### This file is generated from this <a href="02_basic_analysis.ipynb">Notebook</a>

reviews = pd.read_csv("../../data/reviews.csv")
reviews = reviews.drop(columns=["Unnamed: 0"])
reviews.dropna(subset='reviewText', inplace=True)

reviews.head()

reviews.shape

softwares = pd.read_csv("../../data/softwares.csv")
softwares = softwares.drop(columns=["Unnamed: 0"])

reviews = reviews.merge(softwares, on='asin', how='inner')
reviews = reviews[['overall', 'verified', 'reviewTime', 'reviewerID', 'asin','reviewerName', 'reviewText', 'summary', 'vote']]

reviews.info()

softwares.info()

def preprocess_text(text):
    # Tokenize the text
    text = str(text)
    text = text.replace("<div>",'')
    text = text.replace("< div >",'')
    text = text.replace("</div>",'')
    text = text.replace("< /div >",'')
    text = text.replace("< br/ >",'') 
    text = text.replace("< br >",'')
    text = text.replace("< b >",'')
    text = text.replace("< /b >",'')
    text = text.replace("``",'')
    text = text.replace("< strong >",'')
    text = text.replace("< /strong >",'')
    text = text.replace("[",'')
    text = text.replace("]",'')
    tokens = word_tokenize(text.lower())
    # Remove stop words
    filtered_tokens = [token for token in tokens if token not in stopwords.words('english')]
    # Lemmatize the tokens
    lemmatizer = WordNetLemmatizer()
    lemmatized_tokens = [lemmatizer.lemmatize(token) for token in filtered_tokens]
    # Join the tokens back into a string
    processed_text = ' '.join(lemmatized_tokens)
    return processed_text

softwares['title'] = softwares['title'].apply(preprocess_text)
reviews['summary'] = reviews['summary'].apply(preprocess_text)
softwares['description'] = softwares['description'].apply(preprocess_text) 
reviews['reviewText'] = reviews['reviewText'].apply(preprocess_text)
softwares['category'] = softwares['category'].apply(preprocess_text)

### Generate Embeddings for text features

name = 'bert-base-nli-mean-tokens'
model = SentenceTransformer(name)

def get_embeddings(text):
    sentences = list(text)
    return model.encode(sentences)

<br> Realized that the description and asin were unique for each software, hence we have only 650 distinct softwares.

reviews = reviews.merge(softwares, on='asin', how='inner')
reviews.dropna(subset='reviewText', inplace=True)

reviews.head()

from transformers import pipeline
# Output labels
candidate_labels = ["positive", "negative"]
# device=0 for GPU usage
classifier = pipeline("zero-shot-classification",
                      model="facebook/bart-large-mnli", device=0)


def compute_total_sentiment_scores(df):
    postive_score = 0
    negative_score = 0
    for index, row in df.iterrows():
        input_text = row['reviewText']

        model_dict = classifier(input_text, candidate_labels, multi_label=True)

        # Zip results to dict
        result_dict = dict(zip(model_dict.get('labels'), model_dict.get('scores')))
        # get confidence scores
        postive_score += result_dict.get('positive')
        negative_score += result_dict.get('negative')
    nos_records = df.shape[0]
    return postive_score, negative_score, nos_records

unique_ids = reviews.asin.unique().tolist()
scores_dict = {
"asin": [],
"postive_score":[],
"negative_score": [],
"number_reviews": []
}
for id in unique_ids:
    postive_score, negative_score, len_df = compute_total_sentiment_scores(reviews[reviews.asin == id])
    scores_dict['asin'].append(id)
    scores_dict['postive_score'].append(postive_score)
    scores_dict['negative_score'].append(negative_score)
    scores_dict['number_reviews'].append(len_df)
    
scores_df = pd.DataFrame.from_dict(scores_dict)
scores_df.head()

# software_data = review_meta.groupby(['title']).agg({'asin':"first",'description':'first','price':'first','Licensing_Fee':'first','Implemention_cost':'first','Maintenance_cost':'first'}).reset_index()
# software_data['description'] = software_data['description'].apply(preprocess_text)
software_data = software.merge(scores_df, on='asin', how='inner')

software_data.head()

software_data['description_embeddings'] = software_data['description'].apply(lambda x: get_embeddings(x))
software_data['title_embeddings'] = software_data['title'].apply(lambda x: get_embeddings(x))
software_data = software_data[['asin','description_embeddings']]

len(review_meta)

review_meta =  review_meta.merge(software_data, on='asin', how='inner')

review_meta.head()

len(review_meta)

### 

### Correcting ratings with Review Text
We use kmeans to cluster the tfidvector of all reviewtext to correct the ratings. 

#define vectorizer parameters
tfidf_vectorizer = TfidfVectorizer()
tfidf_matrix = tfidf_vectorizer.fit_transform(reviews['reviewText']) #fit the vectorizer to synopses

print(tfidf_matrix.shape)

km = KMeans(5)
km.fit(tfidf_matrix);

ls_val = km.labels_.tolist()

clusters= { 'cluster_1' : ls_val.count(0),
           'cluster_2' : ls_val.count(1),
           'cluster_3' : ls_val.count(2),
           'cluster_4' : ls_val.count(3),
           'cluster_5' : ls_val.count(4),
}

clusters

# Add kmeans clusters
reviews['clusters'] = ls_val

#### For each cluster we count the number of positives and negatives in them the use to encode ratings 1 - 5

# Output labels
candidate_labels = ["positive", "negative"]

# device=0 for GPU usage
classifier = pipeline("zero-shot-classification",
                      model="facebook/bart-large-mnli", device=0)


def count_predictions(df):
    predictions = []
    # Iterate over each row in the DataFrame
    for i in range(len(df)):
        input_text = df['reviewText'].iloc[i]

        # Perform zero-shot classification
        model_dict = classifier(input_text, candidate_labels, multi_label=True)

        # Zip results to dict
        result_dict = dict(zip(model_dict['labels'], model_dict['scores']))

        # Assign the highest scoring label as the predicted label
        predicted_label = max(result_dict, key=result_dict.get)
        predictions.append(predicted_label)

    return {"postive":predictions.count('positive'), "negative":predictions.count('negative')}
    

## For cluster 1
count_predictions(reviews[reviews.clusters == 0])

2662 / (2662 + 55)

## For cluster 2
count_predictions(reviews[reviews.clusters == 1])

447 / (447 + 1)

## For cluster 3
count_predictions(reviews[reviews.clusters == 2])

305 / (305 +5)

## For cluster 4
count_predictions(reviews[reviews.clusters == 3])

3631 / (3631 + 2508)

## For cluster 5
count_predictions(reviews[reviews.clusters == 4])

870 / (870 + 53)

##### We assigned ratings based on the group with the realtive highest positive sentiments

rating_map = {0:2,
 1:5,
 2:4,
 3:1,
 4:3}

review_meta['ratings'] = review_meta['clusters'].map(rating_map)

review_meta.head()

review_meta.description.head()

software_data = pd.read_csv("../../data/softwares_only.csv")
software_data.drop(columns=['Unnamed: 0'],inplace=True)
software_data.head()

def ranking_algol(df):
    df['rank_score'] = (df['postive_score'] - df['negative_score']) * df['number_reviews']
    df = df.sort_values(by='rank_score', ascending=False)
    return df

def price_ranking(max_price, min_price, max_license_price, min_license_price, max_maintenance_price, min_maintenance_price, ranked_data):
    msg = []
    ranked_price_data = ranked_data[(ranked_data.price >= min_price) & (ranked_data.price <= max_price)]
    if len(ranked_price_data) >= 1:
        ranked_data = ranked_price_data
        # put logging here
    else:
        msg.append("The data available does not have price between the range you specified")

    ranked_license_data = ranked_data[(ranked_data.Licensing_Fee >= min_license_price) & (ranked_data.Licensing_Fee <= max_license_price)]
    if len(ranked_license_data) >= 1:
        ranked_data = ranked_license_data
        # put logging here
    else:
        msg.append("The data available does not have license fee between the range you specified after the price filtering.")

    ranked_maintenance_data = ranked_data[(ranked_data.Licensing_Fee >= min_license_price) & (ranked_data.Licensing_Fee <= max_license_price)]
    if len(ranked_maintenance_data) >= 1:
        ranked_data = ranked_maintenance_data
        # put logging here
    else:
        msg.append("The data available does not have the maintenance fee between the range you specified after license fee filtering")

    return ranked_data , msg





def rec_softwares(software_description, software_data, model_name, max_price = np.inf, min_price = -1, max_license_price = np.inf, min_license_price = -1, max_maintenance_price = np.inf, min_maintenance_price = -1):
    software_data['software_description'] = software_description
    
    if model_name == "TfidfVectorizer":
        vectorizer = TfidfVectorizer()
        X = vectorizer.fit_transform(software_data[['software_description','description']])
        cosine_sim = cosine_similarity(X)
    else:
        try:
            model = SentenceTransformer(model_name)
            X_1 = model.encode(software_data['software_description'])
            X_2 = model.encode(software_data['description'])
            cosine_sim = cosine_similarity(X_1, X_2)
        except Exception as e:
            raise f"{model_name} not a transformer model or TfidVectorizer"

    target_item_index = 0 
    scores = list(enumerate(cosine_sim[target_item_index]))
    scores = sorted(scores, key=lambda x: x[1], reverse=True)
    top_n = 10
    top_recommendations = scores[0:top_n+1]
    index_list = []
    for ind, score in top_recommendations:
        index_list.append(ind)
    ranked_data =  ranking_algol(software_data.iloc[index_list,:])
    price_ranked_data, msg = price_ranking(max_price, min_price, max_license_price, min_license_price, max_maintenance_price, min_maintenance_price, ranked_data)

    if msg:
        for m in msg:
            print(m)
    
    if len(price_ranked_data) > 2:
        return price_ranked_data
    else:
        return price_ranked_data


## Experiments

#### Using TfidVectorizer

data = rec_softwares("HR program for windows and mac for japanese humans",software_data,"TfidfVectorizer")

data

#### Recommending based on LLM

model_name = 'sentence-transformers/all-mpnet-base-v2'
rec_softwares("HR program for windows and mac for japanese humans",software_data,model_name)

model_name = 'sentence-transformers/paraphrase-albert-small-v2'
rec_softwares("HR program for windows and mac for japanese humans",software_data,model_name)

model_name = 'sentence-transformers/all-MiniLM-L6-v2'
rec_softwares("HR program for windows and mac for japanese humans",software_data,model_name)
